+++ CHANGELOG.md
@@ -4,6 +4,8 @@ NEW FEATURES:
 
 BUG FIXES:
 
+* Fix per-project prune option handling ([#1562](https://github.com/golang/dep/pull/1562))
+
 IMPROVEMENTS:
 
 # v0.4.0